requirements.txt
Django==3.0.5
django-bootstrap4==1.1.1
django-bootstrap-datepicker-plus==3.0.5
django-crispy-forms==1.9.0
